[fio extras]: add 'length' and 'hashes' to snapshot
Add 'length' and 'hashes' attributes of targets to snapshot metadata.
It is required for a case when different targets have the same version.
If just a version attribute is specified then snapshots of these two
different targets are identical, hence it's not possible for a client
to determine whether to update metadata or not.

Signed-off-by: Mike Sul <mike.sul@foundries.io>